+++ expenses/templatetags/query_params.py
@@ -1,0 +1,18 @@
+from django import template
+from urllib.parse import urlencode
+
+register = template.Library()
+
+
+@register.simple_tag(takes_context=True)
+def query_transform(context, **kwargs):
+    request = context["request"]
+    query_params = request.GET.copy()
+
+    if "page" in query_params:
+        query_params.pop("page")
+
+    for key, value in kwargs.items():
+        query_params[key] = value
+
+    return query_params.urlencode()